refactor(solver): split shared query cache shard
