+++ crates/tsz-solver/src/caches/mod.rs
@@ -5,4 +5,5 @@ mod query_cache_evaluation;
 pub(crate) mod query_cache_statistics;
 pub(crate) mod query_trace;
 mod shared_instantiation;
+pub(crate) mod shared_query_cache;
 pub(crate) mod subtype_reduction_cache;
